scenes resize based on screen size

+++ Assets/Scripts/BioStory.cs
@@ -8,15 +8,11 @@ public class BioStory : MonoBehaviour {
 	public Text btnText;
 	private int countIntro = 0;
 
-	// Use this for initialization
-	void Start () {
-		//setFlubText("Flubba: Huh?");
-	}
-	
 	// Update is called once per frame
 	void Update () {
-		
-		
+		if (Input.GetKeyDown("return")){
+			clickedIntro();
+		}
 	}
 	
 	public void setEvilText(string msg){
@@ -25,27 +21,18 @@ public class BioStory : MonoBehaviour {
 	
 	public void clickedIntro(){
 		if (countIntro == 0) {
-			StartCoroutine ("multipleTexts1");
+			setEvilText ("Voice: Congratulations, you've escaped from Bohr's chemistry lab.");
 			countIntro++;
 		} else if (countIntro == 1) {
-			Application.LoadLevel(3);
+			setEvilText ("Voice: This is Bohr's biology lab where he spends most of his time. You should find him here.");
+			countIntro++;
+		} else if (countIntro == 2) {
+			setEvilText ("Voice: You're going to have to kill him if you want any chance of escaping.");
+			countIntro++;
+		} else if (countIntro == 3) {
+			Application.LoadLevel(1);
 		}
 	}
-	
-	IEnumerator multipleTexts1(){
-		btnText.text = "";
-		setEvilText ("Voice: Congratulations, you've escaped from Bohr's chemistry lab.");
-		//System.Threading.Thread.Sleep (1000);
-		yield return new WaitForSeconds(2);
-		setEvilText ("Voice: This is Bohr's biology lab where he spends most of his time. You should find him here.");
-		//System.Threading.Thread.Sleep (1000);
-		yield return new WaitForSeconds(2);
-
-		setEvilText ("Voice: You're going to have to kill him if you want any chance of escaping.");
-
-		btnText.text = "Continue";
-	}
-
 	
 	IEnumerable changeScene(){
 		float fadeTime = GameObject.Find("IntroCanvas").GetComponent<Fading>().BeginFade(1);
